+++ src/main/scala/write/MilvusBatchWriter.scala
@@ -1,0 +1,30 @@
+package com.zilliz.spark.connector.write
+
+import org.apache.spark.sql.types.StructType
+import org.apache.spark.sql.connector.write.{
+  BatchWrite,
+  DataWriterFactory,
+  PhysicalWriteInfo,
+  Write,
+  WriterCommitMessage
+}
+import com.zilliz.spark.connector.MilvusOption
+
+case class MilvusWrite(milvusOptions: MilvusOption, schema: StructType)
+    extends Write
+    with Serializable {
+  override def toBatch: BatchWrite = MilvusBatchWriter(milvusOptions, schema)
+}
+
+case class MilvusBatchWriter(milvusOptions: MilvusOption, schema: StructType)
+    extends BatchWrite {
+  override def createBatchWriterFactory(
+      info: PhysicalWriteInfo
+  ): DataWriterFactory = {
+    MilvusDataWriterFactory(milvusOptions, schema)
+  }
+
+  override def commit(messages: Array[WriterCommitMessage]): Unit = {}
+
+  override def abort(messages: Array[WriterCommitMessage]): Unit = {}
+}
