+++ src/main/scala/write/MilvusDataWriterFactory.scala
@@ -1,0 +1,19 @@
+package com.zilliz.spark.connector.write
+
+import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.connector.write.{DataWriter, DataWriterFactory}
+import org.apache.spark.sql.types.StructType
+import com.zilliz.spark.connector.MilvusOption
+
+case class MilvusDataWriterFactory(
+    milvusOptions: MilvusOption,
+    schema: StructType
+) extends DataWriterFactory
+    with Serializable {
+  override def createWriter(
+      partitionId: Int,
+      taskId: Long
+  ): DataWriter[InternalRow] = {
+    MilvusInsertDataWriter(partitionId, taskId, milvusOptions, schema)
+  }
+}
